+++ test/index/tocless_test.js
@@ -1,6 +1,6 @@
 const compile = require('../../index')()
 
-describe('compile: toc-less', function () {
+describe('index/toc-less:', function () {
   beforeEach(function (done) {
     // Mock metalsmith object
     this.ms = {
@@ -14,13 +14,16 @@ describe('compile: toc-less', function () {
     }
 
     compile(this.files, this.ms, (err) => {
-      this.err = err
+      if (err) throw err
       done()
     })
   })
 
-  it('fails', function () {
-    expect(this.err).toExist()
-    expect(this.err.message).toMatch(/not found/)
+  it('works', function () {
+    expect(this.files['index.html']).toExist()
+  })
+
+  it('renders title', function () {
+    expect(this.files['index.html'].title).toEqual('hello')
   })
 })